requirements.txt
appdirs==1.4.4
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
distlib==0.3.1
filelock==3.0.12
google-api-core==1.26.3
google-api-python-client==2.4.0
google-auth==1.30.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==2.10
packaging==20.9
protobuf==3.17.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dotenv==0.19.0
pytz==2021.1
requests==2.25.1
rsa==4.7.2
six==1.16.0
uritemplate==3.0.1
urllib3==1.26.4
virtualenv==20.4.6
